_template-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
yt-dlp
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
jinja2==3.1.4
python-multipart==0.0.12
